requirements.txt
discord.py
Pillow
pytz